% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criterion.R
\name{criterion}
\alias{criterion}
\title{Compute the empirical L_2 distance related to the warping function.}
\usage{
criterion(t, f, g, theta, splineBasisW)
}
\arguments{
\item{t}{A vector of numbers, corresponding to time points.}

\item{f}{A vector of numbers, corresponding to the evaluated function.}

\item{g}{A vector of numbers, corresponding to the evaluated function.}

\item{theta}{A vector of warping parameters.}

\item{splineBasisW}{A matrix, corresponding to the spline basis for
the warping functions, evaluted in time points.}
}
\value{
A list, with crit the distance.
}
\description{
This function returns the empirical L_2 between two functions, the
first one being warped.
}

