% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimationTheta.R
\name{estimationTheta}
\alias{estimationTheta}
\title{Estimate the warping parameters.}
\usage{
estimationTheta(t, y, splineBasisW, indSignal, thetaObs)
}
\arguments{
\item{t}{A vector of numbers, corresponding to time points.}

\item{y}{A matrix of numbers, corresponding to observations (size: T * n).}

\item{splineBasisW}{A matrix, corresponding to the spline basis for
the warping functions, evaluted in time points.}

\item{indSignal}{A matrix, corresponding to the individual aligned curves.}

\item{thetaObs}{A matrix, corresponding to initial values for the warping parameters.}
}
\value{
A list, with theta, a matrix of estimated warping parameters,
        and wT, the corresponding warping functions.
}
\description{
This function estimate the warping parameters, knowing the observations and
the individual aligned curves.
}

