% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialisationPara.R
\name{initialisationPara}
\alias{initialisationPara}
\title{Initialize the functional parameters (associated to the aligned curves).}
\usage{
initialisationPara(t, y, splineBasisMu, splineBasisU, warpTime)
}
\arguments{
\item{t}{A vector of numbers, corresponding to time points.}

\item{y}{A matrix of numbers, corresponding to observations (size: T * n).}

\item{splineBasisMu}{A matrix, corresponding to the spline basis for
the global mean function, evaluted in time points.}

\item{splineBasisU}{A matrix, corresponding to the spline basis for
the individual curves, evaluted in time points.}

\item{warpTime}{A matrix, corresponding to warping time points.}
}
\value{
A list, with x, aligned curves, alphaMu the coefficients of the mean curve,
        sigmaEpsilon the variance of the noise, sigmaU the variance of the random effects,
        and indSignal each individual curves.
}
\description{
This function initializes the mean curve, the individual effect U_i, related to
aligned curves.
}

