% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialisationTheta.R
\name{initialisationTheta}
\alias{initialisationTheta}
\title{Initialize the warping parameters.}
\usage{
initialisationTheta(t, y, splineBasisW)
}
\arguments{
\item{t}{A vector of numbers, corresponding to time points.}

\item{y}{A matrix of numbers, corresponding to observations (size: T * n).}

\item{splineBasisW}{A matrix, corresponding to the spline basis for
the warping functions, evaluted in time points.}
}
\value{
A list, with theta, a matrix of estimated warping parameters,
        and wT, the corresponding warping functions.
}
\description{
This function initializes the warping parameters
}

