% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/majPara.R
\name{majPara}
\alias{majPara}
\title{Update the functional parameters (associated to the aligned curves).}
\usage{
majPara(t, y, splineBasisMu, splineBasisU, warpTime)
}
\arguments{
\item{t}{A vector of numbers, corresponding to time points.}

\item{y}{A matrix of numbers, corresponding to observations (size: T * n).}

\item{splineBasisMu}{A matrix, corresponding to the spline basis for
the global mean function, evaluted in time points.}

\item{splineBasisU}{A matrix, corresponding to the spline basis for
the individual curves, evaluted in time points.}

\item{warpTime}{A matrix, corresponding to warping time points.}
}
\value{
A list, with x, aligned curves, alphaMu the coefficients of the mean curve,
        sigmaEpsilon the variance of the noise, sigmaU the variance of the random effects,
        and indSignal the individual curves.
}
\description{
This function updates the estimations of the mean curve,
the individual effect U_i, related to aligned curves.
}

