% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictionTheta.R
\name{predictionTheta}
\alias{predictionTheta}
\title{Predict the warping parameters.}
\usage{
predictionTheta(thetaObs, sigmaEpsilon)
}
\arguments{
\item{thetaObs}{A matrix (size: n * T) corresponding of the estimations of the warping parameters.}

\item{sigmaEpsilon}{A number, defining the variance of the noise in the linear mixed-
effect model fitted on the warping parameters.}
}
\value{
A list, with theta, a matrix of predicted warping parameters,
        sigmaE the covariance of the random effects, and theta0 the mean.
}
\description{
This function predict the warping parameters, using the estimations of those parameters,
and fitting a linear mixed effect model on them.
}

