% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warpTimeFunction.R
\name{warpTimeFunction}
\alias{warpTimeFunction}
\title{Compute the warped time points}
\usage{
warpTimeFunction(splineBasisW, theta, t)
}
\arguments{
\item{splineBasisW}{A matrix, corresponding to the spline basis for
the warping functions, evaluted in time points.}

\item{theta}{A matrix, corresponding to initial values for the warping parameters.}

\item{t}{A vector of numbers, corresponding to time points.}
}
\value{
A vector, corresponding to the warped time points.
}
\description{
This function returns warped time points for a known warping parameter theta.
}

